Fixed a little bug and a minor problem in the final twist behavior. Added functionality to keep commanded velocity until the region near the goal point is reached (Now, PD control on the position error takes only place in this region, else commanded speed is applied.). Did a little refactoring.

+++ include/vehicle_controller/utility.h
@@ -27,7 +27,7 @@ inline static double angularNorm(double diff)
     return diff;
 }
 
-inline static float euclideanDistance(geometry_msgs::Point const & p0, geometry_msgs::Point const & p1)
+inline static double euclideanDistance(geometry_msgs::Point const & p0, geometry_msgs::Point const & p1)
 {
   return std::sqrt(std::pow(p1.x - p0.x, 2) + std::pow(p1.y - p0.y, 2) + pow(p1.z - p0.z, 2));
 }
